(2) DefaultAsyncJobExecutor refactoring.
